requirements.txt
flask
torch
transformers
diffusers
accelerate
safetensors
Pillow